migrate pkgrepo tests to test_debian.py
