+++ tests/pytests/functional/states/pkgrepo/test_debian.py
@@ -507,3 +507,71 @@ def test_repo_present_absent_trailing_slash_uri(pkgrepo, trailing_slash_repo_fil
     # absent
     ret = pkgrepo.absent(name=repo_content)
     assert ret.result
+
+
+@pytest.fixture
+def key_path():
+    key_file = pathlib.Path("/usr", "share", "keyrings", "salt-archive-keyring.gpg")
+    assert not key_file.is_file()
+    yield key_file
+    key_file.unlink()
+
+
+def test_adding_repo_file_signedby(grains, states, tmp_path, key_path):
+    """
+    Test adding a repo file using pkgrepo.managed
+    and setting signedby
+    """
+    repo_file = str(tmp_path / "stable-binary.list")
+    fullname = grains["osfullname"].lower().split()[0]
+    arch = grains["osarch"]
+    lsb_release = grains["lsb_distrib_release"]
+    key_file = "https://repo.saltproject.io/py3/{}/{}/{}/latest/salt-archive-keyring.gpg".format(
+        fullname, lsb_release, arch
+    )
+    repo_content = "deb [arch={arch} signed-by=/usr/share/keyrings/salt-archive-keyring.gpg] https://repo.saltproject.io/py3/{}/{}/{arch}/latest {} main".format(
+        fullname, lsb_release, grains["oscodename"], arch=arch
+    )
+    ret = states.pkgrepo.managed(
+        name=repo_content,
+        file=repo_file,
+        clean_file=True,
+        signedby=str(key_path),
+        key_url=key_file,
+        aptkey=False,
+    )
+    with salt.utils.files.fopen(repo_file, "r") as fp:
+        file_content = fp.read()
+        assert file_content.strip() == repo_content
+    assert key_path.is_file()
+
+
+def test_adding_repo_file_signedby_keyserver(grains, states, tmp_path, key_path):
+    """
+    Test adding a repo file using pkgrepo.managed
+    and setting signedby with a keyserver
+    """
+    repo_file = str(tmp_path / "stable-binary.list")
+    fullname = grains["osfullname"].lower().split()[0]
+    arch = grains["osarch"]
+    lsb_release = grains["lsb_distrib_release"]
+    key_file = "https://repo.saltproject.io/py3/{}/{}/{}/latest/salt-archive-keyring.gpg".format(
+        fullname, lsb_release, arch
+    )
+    repo_content = "deb [arch={arch} signed-by=/usr/share/keyrings/salt-archive-keyring.gpg] https://repo.saltproject.io/py3/{}/{}/{arch}/latest {} main".format(
+        fullname, lsb_release, grains["oscodename"], arch=arch
+    )
+
+    ret = states.pkgrepo.managed(
+        name=repo_content,
+        file=repo_file,
+        clean_file=True,
+        signedby=str(key_path),
+        keyserver="keyserver.ubuntu.com",
+        keyid="0E08A149DE57BFBE",
+        aptkey=False,
+    )
+    with salt.utils.files.fopen(repo_file, "r") as fp:
+        file_content = fp.read()
+        assert file_content.strip() == repo_content
+    assert key_path.is_file()
